Main pages fixed container-component structure
